# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MenuVersion {
  version: number;
  timestamp: string;
}

export interface RestaurantSettings {[key: string]: string}

export interface Category {
  id: string;
  nameEn: string;
  nameAr: string;
  icon: string;
  displayOrder: number;
}

export interface CategoryInput {
  id?: string;
  nameEn?: string;
  nameAr?: string;
  icon?: string;
  displayOrder?: number;
}

export interface Dish {
  id: string;
  categoryId: string;
  nameEn: string;
  nameAr: string;
  descEn: string;
  descAr: string;
  price: number;
  image: string;
  isChefSpecial: boolean;
  isBestSeller: boolean;
  isVegetarian: boolean;
  isSpicy: boolean;
  inStock: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface DishInput {
  id?: string;
  categoryId?: string;
  nameEn?: string;
  nameAr?: string;
  descEn?: string;
  descAr?: string;
  price?: number;
  image?: string;
  isChefSpecial?: boolean;
  isBestSeller?: boolean;
  isVegetarian?: boolean;
  isSpicy?: boolean;
  inStock?: boolean;
  createdAt?: string;
  updatedAt?: string;
}

export interface SettingsInput {[key: string]: string}

export interface UploadInput {
  base64Data: string;
  filename?: string;
}

export interface UploadResponse {
  success: boolean;
  url: string;
}

export interface MenuResponse {
  settings: RestaurantSettings;
  categories: Category[];
  dishes: Dish[];
  updatedAt: string;
}

export interface BackupResponse {
  exportedAt: string;
  restaurant: string;
  settings: RestaurantSettings;
  categories: Category[];
  dishes: Dish[];
}

export interface RestoreInput {
  settings?: RestaurantSettings;
  categories?: CategoryInput[];
  dishes?: DishInput[];
}

export interface MutationResponse {
  success: boolean;
  id?: string;
  image?: string;
  message?: string;
}