# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Category {
  id: string;
  nameEn: string;
  nameAr: string;
  icon: string;
  displayOrder: number;
}