# lib/api-zod/src/generated/types/dishInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DishInput {
  id?: string;
  categoryId?: string;
  nameEn?: string;
  nameAr?: string;
  descEn?: string;
  descAr?: string;
  price?: number;
  image?: string;
  isChefSpecial?: boolean;
  isBestSeller?: boolean;
  isVegetarian?: boolean;
  isSpicy?: boolean;
  inStock?: boolean;
  createdAt?: string;
  updatedAt?: string;
}