# lib/api-zod/src/generated/types/menuResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from './category';
import type { Dish } from './dish';
import type { RestaurantSettings } from './restaurantSettings';

export interface MenuResponse {
  settings: RestaurantSettings;
  categories: Category[];
  dishes: Dish[];
  updatedAt: string;
}