# lib/api-zod/src/generated/types/mutationResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MutationResponse {
  success: boolean;
  id?: string;
  image?: string;
  message?: string;
}