# lib/api-zod/src/generated/types/restoreInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryInput } from './categoryInput';
import type { DishInput } from './dishInput';
import type { RestaurantSettings } from './restaurantSettings';

export interface RestoreInput {
  settings?: RestaurantSettings;
  categories?: CategoryInput[];
  dishes?: DishInput[];
}